Adding post header component

+++ client/src/api/UserRequest.js
@@ -9,5 +9,4 @@ export const getAllUser = () => API.get('/user');
 
 export const followUser = (id, data) => API.put(`/user/${id}/follow`, data);
 export const unfollowUser = (id, data) => API.put(`/user/${id}/unfollow`, data);
-export const deleteUser = (id, userId, isAdmin) =>
-  API.delete(`/user/${id}`, isAdmin);
+export const deleteUser = (id, isAdmin) => API.delete(`/user/${id}`, isAdmin);
